Move IoSourceState logic to selector files
The old macro was a mess of not(any(targets...)), means that when
porting to a new target you always used the stateless implementation
even if you didn't know about it.

This fixes gotcha that and moves the logic of what kind of IoSourceState
to use to the selector implemtnation files, which means that we don't
have to worry about it when porting to a new platform.
